feat(clone): refactor of snapshot and clone code
