+++ io-engine/src/grpc/v1/replica.rs
@@ -100,6 +100,7 @@ impl From<Lvol> for Replica {
             poolname: l.pool_name(),
             usage: Some(usage.into()),
             allowed_hosts: l.allowed_hosts(),
+            is_snapshot: l.is_snapshot(),
             is_clone: l.is_clone(),
             snapshot_uuid: source_uuid,
         }
